requirements.txt
pandas==2.2.0
mysql-connector-python==9.0.0